Create server-side compatibility package
