+++ e2e/bare/driver.py
@@ -2,7 +2,7 @@ import flwr as fl
 
 
 # Start Flower server
-hist = fl.server.driver.start_driver(
+hist = fl.server.start_driver(
     server_address="0.0.0.0:9091",
     config=fl.server.ServerConfig(num_rounds=3),
 )
