+++ src/py/flwr/server/__init__.py
@@ -18,12 +18,13 @@
 from . import driver, strategy
 from .app import run_driver_api as run_driver_api
 from .app import run_fleet_api as run_fleet_api
-from .app import run_server_app as run_server_app
 from .app import run_superlink as run_superlink
 from .app import start_server as start_server
 from .client_manager import ClientManager as ClientManager
 from .client_manager import SimpleClientManager as SimpleClientManager
+from .compat import start_driver as start_driver
 from .history import History as History
+from .run_serverapp import run_server_app as run_server_app
 from .server import Server as Server
 from .server_config import ServerConfig as ServerConfig
 from .serverapp import ServerApp as ServerApp
@@ -40,6 +41,7 @@ __all__ = [
     "ServerApp",
     "ServerConfig",
     "SimpleClientManager",
+    "start_driver",
     "start_server",
     "strategy",
 ]
