+++ src/py/flwr/server/compat/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Flower Labs GmbH. All Rights Reserved.
+# Copyright 2024 Flower Labs GmbH. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,13 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower driver SDK."""
+"""Flower ServerApp compatibility package."""
 
 
-from .driver import Driver
-from .grpc_driver import GrpcDriver
+from .app import start_driver as start_driver
 
 __all__ = [
-    "Driver",
-    "GrpcDriver",
+    "start_driver",
 ]
